refactor: decompose component/Homepage
