+++ src/components/homepage/NewSectionCreator.jsx
@@ -1,0 +1,28 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import elementStyles from '../../styles/isomer-cms/Elements.module.scss';
+
+const NewSectionCreator = ({ createHandler, hasResources }) => (
+  <div className={`${elementStyles.card} ${elementStyles.addNew}`}>
+    <h2>
+Add a new section
+      <i className="bx bx-plus" />
+    </h2>
+    <select name="newSection" id="section-new" onChange={createHandler}>
+      <option value="">--Please choose a new section--</option>
+      <option value="infobar">Infobar</option>
+      {/* If homepage already has a Resources section,
+        don't display the option to create one */}
+      {hasResources
+        ? null
+        : <option value="resources">Resources</option>}
+    </select>
+  </div>
+);
+
+export default NewSectionCreator;
+
+NewSectionCreator.propTypes = {
+  createHandler: PropTypes.func.isRequired,
+  hasResources: PropTypes.bool.isRequired,
+};
